// src/types/Employee.ts
export interface Employee {
  rol: string;

  firstName: string;
  lastName: string;
  phoneCard: string;
  email: string;
  website: string;
  fullName: string;
  profilePhotoUrl: string;
  whspUrl: string;
  facebookUrl: string;
  emailUrl: string;
  instagramUrl: string;
  phone: string;
  youtubeUrl: string;
}